UserManagement/router.py
from django.urls import path
from . import views
from rest_framework.routers import SimpleRouter, DefaultRouter

router = DefaultRouter()
router.register(r'', views.UserViewSet)
urlpatterns = router.urls
