requirements.txt
certifi==2018.10.15
chardet==3.0.4
Django==2.1.3
django-webpack-loader==0.6.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
idna==2.7
pkg-resources==0.0.0
PyJWT==1.6.4
pytz==2018.7
requests==2.20.0
six==1.11.0
urllib3==1.24.1
